# backend/src/main/resources/swagger/notice-board.yaml
openapi: 3.0.0
info:
  title: Ourhouse NoticeBoard API
  description: <h2>우리의집 공지사항 API</h2>
  version: 1.0.0
security:
  - bearerAuth: []
#servers:
#  - url: https://devdibi.site/ousrhouse/
#    description: ourhouse 서비스
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      scheme: bearer
      bearerFormat: JWT
      name: Authorization
      in: header
paths:
  # 유저 - 로그인
  /user/sign-in:
    post:
      tags:
        - user
      summary: 로그인
      requestBody:
        description: "로그인"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: test_admin@gmail.com
                password:
                  type: string
                  example: test_admin
      responses:
        200:
          description: 로그인 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: OK
                  msg:
                    type: string
                    example: 로그인 완료
                  data:
                    type: string
                    example: token
        404:
          description: 이메일 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: NOT_FOUND
                  msg:
                    type: string
                    example: 존재하지 않는 회원입니다.
                  data:
                    type: string
                    example: null
        400:
          description: 비밀번호 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: BAD_REQUEST
                  msg:
                    type: string
                    example: 잘못된 비밀번호입니다.
                  data:
                    type: string
                    example: null
  # 공지사항 - 게시글 작성
  /notice-board/create:
    post:
      security:
        - bearerAuth: []
      tags:
        - notice-board
      summary: 공지사항 작성
      requestBody:
        description: 공지사항 작성
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 제목입니다.
                content:
                  type: string
                  example: 내용입니다.
      responses:
        200:
          description: 게시글 작성 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: OK
                  msg:
                    type: string
                    example: 성공적으로 게시글을 작성했습니다.
                  data:
                    type: object
                    example: null
        400:
          description: 게시글 작성 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: BAD_REQUEST
                  msg:
                    type: string
                    example: 제목 또는 본문이 입력되지 않았습니다.
                  data:
                    type: object
                    example: null
  # 공지사항 - 게시글 목록 조회
  /notice-board/list:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - notice-board
      summary: 공지사항 목록 조회
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 10
      responses:
        200:
          description: 게시글 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: OK
                  msg:
                    type: string
                    example: 성공적으로 게시글 목록을 조회했습니다.
                  data:
                    type: object
                    example: null
        400:
          description: 게시글 목록 조회 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: BAD_REQUEST
                  msg:
                    type: string
                    example: 목록을 조회하는데 문제가 생겼습니다.
                  data:
                    type: object
                    example: null
  # 공지사항 - 게시글 상세 조회
  /notice-board/read:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - notice-board
      summary: 공지사항 조회
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 1
      responses:
        200:
          description: 게시글 조회 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: OK
                  msg:
                    type: string
                    example: 성공적으로 게시글을 조회했습니다.
                  data:
                    type: object
                    example: { owner: boolean, board: freeBoard }
        400:
          description: 게시글 조회 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: BAD_REQUEST
                  msg:
                    type: string
                    example: 게시글 조회를 실패했습니다.
                  data:
                    type: object
                    example: null
  # 공지사항 - 게시글 수정
  /notice-board/update:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - notice-board
      summary: 공지사항 게시글 수정
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 1
      requestBody:
        description: 공지사항 게시글 수정
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 제목을 수정합니다.
                content:
                  type: string
                  example: 내용을 수정합니다.
      responses:
        200:
          description: 게시글 수정 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: OK
                  msg:
                    type: string
                    example: 정상적으로 게시글을 수정했습니다.
                  data:
                    type: object
                    example: null
        404:
          description: 존재하지 않는 게시글
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: NOT_FOUND
                  msg:
                    type: string
                    example: 게시글이 존재하지 않습니다.
                  data:
                    type: object
                    example: null
        403:
          description: 다른 작성자
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: FORBIDDEN
                  msg:
                    type: string
                    example: 게시글의 작성자가 아닙니다.
                  data:
                    type: object
                    example: null
  # 공지사항 - 게시글 삭제
  /notice-board/delete:
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - notice-board
      summary: 공지사항 삭제
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 1
      responses:
        200:
          description: 게시글 삭제 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: OK
                  msg:
                    type: string
                    example: 정상적으로 게시글을 삭제했습니다.
                  data:
                    type: object
                    example: null
        404:
          description: 존재하지 않는 게시글
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: NOT_FOUND
                  msg:
                    type: string
                    example: 게시글이 존재하지 않습니다.
                  data:
                    type: object
                    example: null
        403:
          description: 다른 작성자
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: FORBIDDEN
                  msg:
                    type: string
                    example: 게시글의 작성자가 아닙니다.
                  data:
                    type: object
                    example: null